move cgen-sql to a new module
