--- cgen-sql/small_example.csv
@@ -1,8 +1,0 @@
-Phrase,Year,MatchCount,VolumeCount
-Auswanderung,1524,1,1
-Auswanderung,1750,2,2
-Auswanderung,1844,1,3
-Auswanderung,1845,1,4
-Auswanderung,1847,1,5
-Auswanderung,1850,4,6
-Auswanderung,1855,5,20
